move every sinlge puts and gets to main.rb
