+++ bin/game.rb
@@ -1,8 +1,6 @@
 # rubocop: disable Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
 
-require_relative 'menu'
 require_relative 'displayprint'
-require_relative 'inputchecker'
 
 class Game
   def initialize(player1, player2)
@@ -14,52 +12,8 @@ class Game
     @display = Display.new
   end
 
-  def take_turns
-    game = 0
-    @current_player = @first_player if @current_player.nil?
-
-    while game.zero?
-      @display.display_clear
-      @display.board_print(@board)
-      puts '+-----------------------------------------------------------+'
-      puts "It is #{@current_player} turn:"
-      puts 'Please select any valid spot or cell on the board to play'
-      input_checker = Inputchecker.new
-      player_asw = input_checker.number_checker(gets.chomp.to_i, 1, 9)
-      player_turns_done(player_asw, @symbol)
-      game = game_ended?
-      next_player if game.zero?
-    end
-    end_game_message(game)
-    end_game
-  end
-
-  def end_game
-    @display.board_print(@board)
-    puts 'Dou you want to start a new game?(Y/N):'
-    puts '1. Yes'
-    puts '2. No'
-    input_checker = Inputchecker.new
-    answ = input_checker.number_checker(gets.chomp.to_i, 1, 2)
-    menu = Menu.new
-    case answ
-    when 1
-      menu.call_game
-    when 2
-      menu.game_menu
-    end
-  end
-
-  def end_game_message(game)
-    @display.display_clear
-    case game
-    when 1
-      @display.jaarix("Congratulations player #{@first_player} you are super!!!!!")
-    when 2
-      @display.jaarix("Congratulations player #{@second_player} you are super!!!!!")
-    when 3
-      @display.jaarix('Its a tie, booo no one won!!!!!')
-    end
+  def only_board
+    @board
   end
 
   def game_ended?
@@ -82,15 +36,6 @@ class Game
     0
   end
 
-  def player_turns_done(asnw, symbol)
-    while @board.none?(asnw)
-      puts 'The location selected is not available, please select another one:'
-      input_checker = Inputchecker.new
-      asnw = input_checker.number_checker(gets.chomp.to_i, 1, 9)
-    end
-    @board[asnw - 1] = symbol
-  end
-
   def next_player
     if @current_player == @first_player
       @current_player = @second_player
@@ -99,6 +44,14 @@ class Game
       @current_player = @first_player
       @symbol = 'X'
     end
+  end
+
+  def p_player
+    @current_player
+  end
+
+  def p_symbol
+    @symbol
   end
 end
