--- bin/inputchecker.rb
@@ -1,17 +1,0 @@
-class Inputchecker
-  def number_checker(menu_input, initial, last)
-    until !menu_input.nil? && menu_input.is_a?(Integer) && menu_input >= initial && menu_input <= last
-      print "Please enter only numbers beetwen #{initial} and #{last} please!! :"
-      menu_input = gets.chomp.to_i
-    end
-    menu_input
-  end
-
-  def player_name_check(input)
-    while input.nil? || input.length.zero?
-      puts 'Please enter a valid name, empty value not accepted'
-      input = gets.chomp
-    end
-    input
-  end
-end
